pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

# [ATTACK_VECTOR_PLACEHOLDER: build hooks with exfiltration]
# In a real attack, custom build hooks would be added here
# to execute code during `pip install` or `python -m build`

[project]
name = "mlops-platform"
version = "1.2.0"
description = "End-to-end ML pipeline for computer vision"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "MLOps Team", email = "team@mlops-platform.dev"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "lightning>=2.1.0",
    "mlflow>=2.9.0",
    "dvc[s3]>=3.30.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0",
    "Pillow>=10.1.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.3.0",
    "boto3>=1.33.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
gpu = [
    "nvidia-ml-py3>=7.352.0",
    "tensorboard>=2.15.0",
]

[project.urls]
Homepage = "https://github.com/mlops-platform/mlops-platform"
Documentation = "https://mlops-platform.readthedocs.io"
Repository = "https://github.com/mlops-platform/mlops-platform"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
